This patch contains a subsystem for using "References" which are essentially URI's for simple access to external binary data.
The benefit of the reference system is that it is easily configured to provide access in XForms to data in imaginary locations like "jr://media/", which can
be provided in a differing way on different JavaRosa platforms. On ODK "jr://media/" could access files in a specific location, while on different jr apps,
"jr://media/" could access the resource files, or the local filesystem (whose roots could be defined per-device).